+++ src/command_modules/azure-cli-iot/azure/cli/command_modules/iot/shared.py
@@ -51,3 +51,13 @@ class UserRole(Enum):
     Admin = 'admin'
     Reader = 'reader'
     Contributer = 'contributer'
+
+
+# pylint: disable=too-few-public-methods
+class RenewKeyType(Enum):
+    """
+    Type of the RegenerateKey for the authorization policy.
+    """
+    Primary = 'primary'
+    Secondary = 'secondary'
+    Swap = 'swap'